# Fix too many SQL parameters during trim sectors