Fixes the following error on `hostd` when `renterd` trims sectors for large contracts:
```
failed to commit contract modifications: transaction failed (attempt 1): failed to trim sectors: failed to delete contract sector roots: failed to delete contract sector roots: too many SQL variables
```

`renterd`'s batch size when trimming is 500000 sectors while the SQLite query parameter limit is around 32000. This change is negligibly slower (~2 microseconds per sector)

**Before**
```
BenchmarkTrimSectors-10           189414             91529 ns/op            189414 sectors          8892 B/op        207 allocs/op
PASS
ok      go.sia.tech/hostd/persist/sqlite        115.901s
```

**After**
```
BenchmarkTrimSectors-10           179298             94208 ns/op            179298 sectors          9768 B/op        252 allocs/op
PASS
ok      go.sia.tech/hostd/persist/sqlite        102.815s
```